Added uservalidation service & remove extra env config
