+++ Backend/config/database.js
@@ -1,5 +1,3 @@
 const mongoose = require("mongoose");
-const dotenv = require("dotenv");
-dotenv.config();
 const uri = process.env.MONGO_DB;
 module.exports = mongoose.connect(uri);
